libmicrohttpd2: Revamp fuzz_connection
